Update 06_Precipitation_individual_sites.py
Taylor does not yet have an answer on what to do with blanks. He's aware that I'm going to be out, so he will reach out to Carolyn once he has had a chance to dig into it further. For now, the direction Taylor gave me was to replace all months with fewer than 25 precip readings with "NaN," instead of zero. 
Changed the count to "greater than" low threshold and "less than or equal to" high threshold
Changed the precision to whole numbers on the count chart. 
Corrected bottom two charts to display as Oct through Nov